fix: use numeric version id for $_SERVER['PHWOOLCON_PHALCON_VERSION']

+++ phwoolcon-package/di.php
@@ -20,7 +20,7 @@ use Phwoolcon\Session;
 use Phwoolcon\Util\Counter;
 use Phwoolcon\View;
 
-$_SERVER['PHWOOLCON_PHALCON_VERSION'] = Version::getId();
+$_SERVER['PHWOOLCON_PHALCON_VERSION'] = (int)Version::getId();
 
 // PHP 7.2: ini_set(): Headers already sent. You cannot change the session module's ini settings at this time
 ini_get('session.use_cookies') and ini_set('session.use_cookies', 0);
